:art: Move ChatActivity
