+++ dEventer/app/src/main/java/es/achraf/deventer/view/fragments/ChatFragment.java
@@ -16,7 +16,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 
-import es.achraf.deventer.ChatActivity;
+import es.achraf.deventer.view.ChatActivity;
 import es.achraf.deventer.R;
 import es.achraf.deventer.model.Event;
 import es.achraf.deventer.view.IView;
